+++ src/main/java/janggi/domain/board/point/DefaultPoint.java
@@ -1,57 +1,15 @@
 package janggi.domain.board.point;
 
 import janggi.domain.board.Direction;
-import java.util.Objects;
 
-public record DefaultPoint(
-        int x,
-        int y
-) implements Point {
+public class DefaultPoint extends Point {
+
+    public DefaultPoint(int x, int y) {
+        super(x, y);
+    }
 
     @Override
     public Point move(Direction direction) {
         throw new IllegalStateException("움직일 수 없습니다.");
-    }
-
-    @Override
-    public boolean isSamePosition(Point point) {
-        return this.x == point.getX() && y == point.getY();
-    }
-
-    @Override
-    public int getX() {
-        return x;
-    }
-
-    @Override
-    public int getY() {
-        return y;
-    }
-
-    @Override
-    public Point copy(Point endPoint) {
-        return new DefaultPoint(endPoint.getX(), endPoint.getY());
-    }
-
-    @Override
-    public boolean isOutOfBoundary() {
-        return x > 10 || x < 1 || y > 9 || y < 1;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        DefaultPoint point = (DefaultPoint) o;
-        return x == point.x && y == point.y;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(x, y);
     }
 }
